backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
boto3==1.29.7
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
datetime